FAC-490: refactor : Move LoadAnswerOptionsByQuestionPort to correct package
